+++ src/Humanizer/Localisation/Tense.cs
@@ -1,8 +1,18 @@
 ﻿namespace Humanizer.Localisation
 {
+    /// <summary>
+    /// Enumerates the possible time references; past or future.
+    /// </summary>
     public enum Tense
     {
+        /// <summary>
+        /// Indicates the future.
+        /// </summary>
         Future,
+
+        /// <summary>
+        /// Indicates the past.
+        /// </summary>
         Past
     }
 }